+++ backend-microservices/user-services/internal/dto/user_dto.go
@@ -7,7 +7,7 @@ import (
 )
 
 type UserDTO struct {
-	ID        uint      `json:"id"`
+	ID        uint64    `json:"id"`
 	FullName  string    `json:"full_name"`
 	Username  string    `json:"username"`
 	Email     string    `json:"email"`
@@ -19,7 +19,7 @@ type UserDTO struct {
 }
 
 type UserResponseDTO struct {
-	ID          uint      `json:"id"`
+	ID          uint64    `json:"id"`
 	FullName    string    `json:"full_name"`
 	Username    string    `json:"username"`
 	Email       string    `json:"email"`
